refactor: ExploreFilters structure and cards layout
